// src/app/_dto/masters/ProdGroupHead.dto.ts
import { ProdGroupDetail } from "../ProdGroupDetail.dto";

export class ProdGroupHead {
    HProdGroup_SysID: number = 0;
    HProdGroup_Code: string = null;
    HProdGroup_Short_Name: string = null;
    HProdGroup_Long_Name: string = null;
    HProdGroup_Group1: number = null;
    HProdGroup_Group2: number = null;
    HProdGroup_Group3: number = null;
    HProdGroup_Group4: number = null;
    HProdGroup_Group5: number = null;
    HProdGroup_Group6: number = null;
    HProdGroup_Group7: number = null;
    HProdGroup_Group8: number = null;
    HProdGroup_Group9: number = null;

    ProdGroupDetail: ProdGroupDetail[] = []

}